# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cynoia API",
    "version": "1.0.0",
    "description": "Backend API for Cynoia - user management system",
    "contact": {
      "name": "Cynoia Team",
      "email": "contact@cynoia.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ConversationParticipant": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer"
          },
          "conversationId": {
            "type": "integer"
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastReadAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "user": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "integer"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              }
            }
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "conversationId": {
            "type": "integer"
          },
          "senderId": {
            "type": "integer"
          },
          "content": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "example": "TEXT"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "sender": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "integer"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              }
            }
          }
        }
      },
      "Conversation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "GROUP"
            ]
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "entitiesId": {
            "type": "integer"
          },
          "createdById": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConversationParticipant"
            }
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/chats": {
      "get": {
        "tags": [
          "Chats"
        ],
        "summary": "List my conversations",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of conversations with last message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Conversation"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chats/{id}/messages": {
      "get": {
        "tags": [
          "Chats"
        ],
        "summary": "List messages in a conversation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chats/private": {
      "post": {
        "tags": [
          "Chats"
        ],
        "summary": "Create or get a private conversation with another user in my entity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "peerUserId"
                ],
                "properties": {
                  "peerUserId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created or existing conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chats/group": {
      "post": {
        "tags": [
          "Chats"
        ],
        "summary": "Create a group conversation within my entity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "memberIds"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "memberIds": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created group conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chats/message": {
      "post": {
        "tags": [
          "Chats"
        ],
        "summary": "Send a message in a conversation I belong to",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "conversationId",
                  "content"
                ],
                "properties": {
                  "conversationId": {
                    "type": "integer"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/equipements": {
      "get": {
        "tags": [
          "Equipements"
        ],
        "summary": "Get all equipements",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of equipements"
          }
        }
      },
      "post": {
        "tags": [
          "Equipements"
        ],
        "summary": "Create a new equipement",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "state": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Equipement created"
          }
        }
      }
    },
    "/api/v1/equipements/entity/{entityId}": {
      "get": {
        "tags": [
          "Equipements"
        ],
        "summary": "Get equipements by entity id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entityId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of equipements"
          }
        }
      }
    },
    "/api/v1/equipements/{id}": {
      "get": {
        "tags": [
          "Equipements"
        ],
        "summary": "Get equipement by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipement object"
          }
        }
      },
      "put": {
        "tags": [
          "Equipements"
        ],
        "summary": "Update an equipement",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Equipement updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Equipements"
        ],
        "summary": "Delete an equipement",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipement deleted"
          }
        }
      }
    },
    "/api/v1/espaces": {
      "get": {
        "tags": [
          "Espaces"
        ],
        "summary": "Get all espaces",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of espaces"
          }
        }
      },
      "post": {
        "tags": [
          "Espaces"
        ],
        "summary": "Create a new espace",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "entityId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Espace created"
          }
        }
      }
    },
    "/api/v1/espaces/entity/{entityId}": {
      "get": {
        "tags": [
          "Espaces"
        ],
        "summary": "Get espaces by entity id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entityId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of espaces"
          }
        }
      }
    },
    "/api/v1/espaces/{id}": {
      "get": {
        "tags": [
          "Espaces"
        ],
        "summary": "Get espace by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Espace object"
          }
        }
      },
      "put": {
        "tags": [
          "Espaces"
        ],
        "summary": "Update an espace",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Espace updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Espaces"
        ],
        "summary": "Delete an espace",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Espace deleted"
          }
        }
      }
    },
...
  },
...
}